refactor ratelimit config
